CAT-13 type revolution reference token as Compact or Wide
CatiaConsolidatedRevolution.reference_token is Compact (0x08) or Wide
(0x0a). CADIR JSON still emits the token byte. The post-hoc 0x08|0x0a
check is deleted.

Checks: cargo fmt --all && cargo check -q --workspace --tests
Hook-Bypass-Reason: the mission requires --no-verify for each non-final item; the full gate runs once at the end.

+++ crates/cadmpeg-codec-catia/src/families/b2/tests.rs
@@ -918,7 +918,7 @@ fn b2_revolution_parser_reads_axis_profile_bounds_and_exact_scale_relations() {
         let records = crate::families::b2::records::b2_revolutions(&stream);
         assert_eq!(records.len(), 1);
         assert_eq!(records[0].pos, 0);
-        assert_eq!(records[0].reference_token, reference_token);
+        assert_eq!(u8::from(records[0].reference_token), reference_token);
         assert_eq!(records[0].profile_allocation_id, 0x1234);
         assert_eq!(records[0].origin, [1.0, 2.0, 3.0]);
         assert_eq!(records[0].direction_x, [1.0, 0.0, 0.0]);
